+++ package.json
@@ -10,8 +10,9 @@
   },
   "license": "MIT",
   "dependencies": {
-    "express": "^4.16.3",
     "lodash.debounce": "^4.0.8",
+    "micro": "^9.1.4",
+    "micro-route": "^2.5.0",
     "mitt": "^1.1.3",
     "next": "^5.1.0",
     "nprogress": "^0.2.0",
@@ -52,7 +53,7 @@
     ]
   },
   "scripts": {
-    "dev": "node server.js",
+    "dev": "node .",
     "build": "next build",
     "start": "next start",
     "precommit": "lint-staged"
